deprecate and simplify the batch prediction script
