+++ applications/DynaCLR/evaluation/linear_classifiers/generate_train_config.py
@@ -9,7 +9,8 @@ datasets that have both.
 from pathlib import Path
 
 import yaml
-from dataset_discovery import (
+
+from utils import (
     CHANNELS,
     TASKS,
     build_registry,
